+++ src/Commands/Admin/RequestPersonalSite.cs
@@ -1,0 +1,41 @@
+﻿using Microsoft.Online.SharePoint.TenantAdministration;
+using Microsoft.Online.SharePoint.TenantManagement;
+using Microsoft.SharePoint.Client;
+
+using PnP.PowerShell.Commands.Base;
+using System;
+using System.Management.Automation;
+using System.Threading;
+
+namespace PnP.PowerShell.Commands.Admin
+{
+    [Cmdlet(VerbsLifecycle.Request, "PersonalSite")]
+    public class RequestPersonalSite : PnPAdminCmdlet
+    {
+        [Parameter(Mandatory = true, ValueFromPipeline = true)]
+        [ValidateNotNull]
+        [ValidateCount(1, 200)]
+        public string[] UserEmails;
+
+        [Parameter(Mandatory = false)]
+        public SwitchParameter NoWait;
+
+        protected override void ExecuteCmdlet()
+        {
+            foreach (var email in UserEmails)
+            {
+                if (string.IsNullOrEmpty(email))
+                {
+                    throw new PSArgumentException("UserEmails contains an empty value");
+                }
+            }
+            var operation = this.Tenant.RequestPersonalSites(UserEmails);
+            ClientContext.Load(operation);
+            ClientContext.ExecuteQueryRetry();
+            if (NoWait.IsPresent)
+            {
+                PollOperation(operation);
+            }
+        }
+    }
+}